+++ read_structure_step/write_structure.py
@@ -13,6 +13,7 @@ directory, and is used for all normal output from this step.
 
 import logging
 from pathlib import PurePath
+import textwrap
 
 import read_structure_step
 from .write import write
@@ -78,34 +79,76 @@ class WriteStructure(seamm.Node):
         if not P:
             P = self.parameters.values_to_dict()
 
-        text = f"Write structure to {P['file']}. "
+        structures = P["structures"]
+        configs = P["configurations"]
+        ignore_missing = P["ignore missing"]
+        if isinstance(ignore_missing, bool):
+            if ignore_missing:
+                ignore_missing = "yes"
+            else:
+                ignore_missing = "no"
 
-        # # What type of file?
-        # extension = ""
-        # filename = P["file"].strip()
-        # file_type = P["file type"]
+        n_per_file = P["number per file"]
 
-        # if self.is_expr(filename) or self.is_expr(file_type):
-        #     extension = "all"
-        # else:
-        #     if file_type != "from extension":
-        #         extension = file_type.split()[0]
-        #     else:
-        #         if filename != "":
-        #             path = PurePath(filename)
-        #             extension = path.suffix
-        #             if extension == ".gz":
-        #                 extension = path.stem.suffix
+        if structures == "current configuration":
+            text = f"The current configuration will be written to {P['file']}. "
+        elif structures == "current system":
+            if configs == "all":
+                text = (
+                    "All the configurations of the current system will be written to "
+                    f"{P['file']}. "
+                )
+            else:
+                text = (
+                    f"Configuration '{configs}' of the current system will be written "
+                    f"to {P['file']}. "
+                )
+            if n_per_file != "all":
+                text += (
+                    " The output will be broken into multiple files containing no "
+                    f"more than {n_per_file} structures each."
+                )
+        elif structures == "all systems":
+            if configs == "all":
+                text = (
+                    "All the configurations of all the systems will be written to "
+                    f"{P['file']}. "
+                )
+            else:
+                text = (
+                    f"Configuration '{configs}' of all systems will be written "
+                    f"to {P['file']}. "
+                )
+                if ignore_missing == "yes":
+                    pass
+                elif ignore_missing == "no":
+                    text += (
+                        "It will be an error if a system does not have the named "
+                        "configuration."
+                    )
+                else:
+                    text += (
+                        f"The value of {ignore_missing} will determine whether it is"
+                        " an error if a system does not have the named configuration."
+                    )
+            if n_per_file != "all":
+                text += (
+                    " The output will be broken into multiple files containing no "
+                    f"more than {n_per_file} structures each."
+                )
+        else:
+            text = (
+                f"The variable {structures} will determine which systems to write out, "
+                f"and {configs} the configurations for those systems."
+            )
+            if n_per_file != "all":
+                text += (
+                    " The output will be broken into multiple files containing no "
+                    f"more than {n_per_file} structures each."
+                )
 
-        # # Get the metadata for the format
-        # metadata = get_format_metadata(extension)
-
-        # if extension == "all" or not metadata["single_structure"]:
-        #   text += seamm.standard_parameters.multiple_structure_handling_description(P)
-        # else:
-        #     text += seamm.standard_parameters.structure_handling_description(P)
-
-        return text
+        text = textwrap.fill(text, initial_indent=4 * " ", subsequent_indent=4 * " ")
+        return self.header + "\n" + text
 
     def run(self):
         """Run a Write Structure step."""
@@ -134,7 +177,7 @@ class WriteStructure(seamm.Node):
             )
 
         # Print what we are doing
-        printer.important(__(self.description_text(P), indent=4 * " "))
+        printer.important(self.description_text(P))
 
         # Write the file into the system
         system_db = self.get_variable("_system_db")
@@ -145,16 +188,19 @@ class WriteStructure(seamm.Node):
         errors = not P["ignore missing"]
         configurations = []
         if structures == "current configuration":
+            n_systems = 1
             configurations.append(configuration)
         elif structures == "current system":
+            n_systems = 1
             if configs == "all":
                 for configuration in system.configurations:
                     configurations.append(configuration)
-                else:
-                    cid = system.get_configuration_id(configs, errors=errors)
-                    if cid is not None:
-                        configurations.append(system.get_configuration(cid))
+            else:
+                cid = system.get_configuration_id(configs, errors=errors)
+                if cid is not None:
+                    configurations.append(system.get_configuration(cid))
         elif structures == "all systems":
+            n_systems = system_db.n_systems
             if configs == "all":
                 for system in system_db.systems:
                     for configuration in system.configurations:
@@ -167,6 +213,14 @@ class WriteStructure(seamm.Node):
 
         n_per_file = P["number per file"]
         n_configurations = len(configurations)
+        if n_configurations > 1:
+            printer.important(
+                __(
+                    f"\n    Writing out {n_configurations} structures from {n_systems} "
+                    "systems.",
+                    indent=4 * " ",
+                )
+            )
         if n_per_file == "all" or n_configurations <= n_per_file:
             write(
                 filename,